+++ CHANGELOG.markdown
@@ -35,6 +35,7 @@
   same for `Clown`.
 * Make the `Functor` instance for `Coyoneda` do the obvious `Coyoneda`
   thing.
+* Add Template Haskell `Lift` instances for all the types other than `Day`.
 
 5.5.11 [2021.04.30]
 -------------------